Overview and Summary
Fix build issues and update the container for cray-generators to
use alpine 3.12.  This prepares this tool to be published to
arti instead of dtr.

Testing

I was able to use this new version of the tool to run

    craypc generators generate cray-service --sections=kubernetes

against a test repository and get a reasonable looking basic
chart.

Risks

This tool is barely used anymore, so the risks are small.